refactor: split modules into differente folders
