+++ src/components/layout/layout.module.scss
@@ -1,0 +1,4 @@
+.container {
+  max-width: 1170px;
+  margin: 0 auto;
+}
